Insert tallow rule in first position
The purpose of tallow is to ban specific hosts. This ban should be
enacted as early as possible in the firwall, which is not achieved by
using "iptables -A", which *appends* at the end of the chain.

This patch ensures that the tallow rule is *inserted* at the top of the
INPUT chain, also matching the behavior implemented for firewall-cmd.

Fixes: #22 ("iptables rules order")