Allow filter list to be an array of strings.
While testing on Fedora 40+ I noticed that sshd messages came in at both
syslog identified `sshd` and `sshd-session`. This was causing the filter
to deliver effectively no login events at all from sshd.

It will be extremely difficult to try and map the exact "one" filter
that works across versions and distributions. A much simpler soluition
is to allow multiple filters. This doesn't interfere in any way, and we
will only receive journal content once.

So from this change on, `filter` can be an `[ "array", "of", "strings"
].

Signed-off-by: Auke Kok <sofar@foo-projects.org>